Restructure in preparation for custom block rewrite
