+++ server/paper/src/main/kotlin/net/stellarica/server/StellaricaServer.kt
@@ -13,10 +13,10 @@ import net.stellarica.server.crafts.starships.commands.StarshipDebugCommands
 import net.stellarica.server.crafts.starships.listeners.InterfaceListener
 import net.stellarica.server.customblocks.CustomBlockListener
 import net.stellarica.server.customblocks.MushroomEventListener
-import net.stellarica.server.customitems.CustomItems
-import net.stellarica.server.customitems.commands.CustomItemCommands
-import net.stellarica.server.customitems.listeners.ItemEnchantListener
-import net.stellarica.server.customitems.listeners.PowerItemBreakListener
+import net.stellarica.server.material.custom.item.CustomItems
+import net.stellarica.server.material.custom.item.commands.CustomItemCommands
+import net.stellarica.server.material.custom.item.listeners.ItemEnchantListener
+import net.stellarica.server.material.custom.item.listeners.PowerItemBreakListener
 import net.stellarica.server.material.block.BlockType
 import net.stellarica.server.multiblocks.MultiblockCommands
 import net.stellarica.server.multiblocks.MultiblockHandler
