+++ server/paper/src/main/kotlin/net/stellarica/server/material/custom/block/CustomBlockListener.kt
@@ -1,4 +1,4 @@
-package net.stellarica.server.customblocks
+package net.stellarica.server.material.custom.block
 
 import net.stellarica.server.utils.extensions.customItem
 import org.bukkit.block.data.MultipleFacing
