--- server/paper/src/main/kotlin/net/stellarica/server/material/item/CustomItemType.kt
@@ -1,4 +1,0 @@
-package net.stellarica.server.material.item
-
-class CustomItemType {
-}
