+++ server/paper/src/main/kotlin/net/stellarica/server/material/type/block/CustomBlockType.kt
