+++ server/paper/src/main/kotlin/net/stellarica/server/material/type/item/ItemType.kt
@@ -1,4 +1,4 @@
-package net.stellarica.server.material.item
+package net.stellarica.server.material.type.item
 
 import org.bukkit.inventory.ItemStack
