+++ server/paper/src/main/kotlin/net/stellarica/server/material/type/item/VanillaItemType.kt
@@ -1,0 +1,4 @@
+package net.stellarica.server.material.type.item
+
+class VanillaItemType {
+}
